Fix Directories not showing up
As it was before, the script given to the `sh` call
for the `find -exec` option was removing not only the
pass' directory (intended behavior), but also the
rest of the directories, acting as if `basename` was
called (unintended behavior).

Now the `${PREFIX}` variable is passed to the
aforementioned `sh` call and with it, it is
possible to do string deletion to achieve the name
of the entry, i. e., its directories and basename
without pass' directory.