Fix lack of action in specific case.
When the user sets the variable
${PASSWORD_STORE_XMENU_DEFAULT_ACTION} to "prompt"
and does not add a "%" action before a field
("@"), the program would avoid all options in
the case statement inside the elif of the last for
loop.

Fix it by checking, when already inside the elif,
if the variable ${ACTION} is "prompt"; if so,
prompt the user to pick either "paste",
"paste-term" or "type".

Although code-wise the location of such a check
it's very odd, its justification comes with the
avoidance of breaking the user flow: if such a
check is done before the iteration the selected
macro, a user who has added a "%" action to it
will be prompted for a useless choice.