Change clip action to use pass' clip function
How the clip action was implemented before was by
piping the stdout of a printf call to xclip. By
doing this, not only the previous content of the x
selection were lost, but also the used selection
was not cleared after the time in
"${PASSWORD_STORE_CLIP_TIME}".

Change the clip action to make use pass' clip
function, so that now both the previous content
are not lost and the clearing time is respected.