Change "${passfile}" to "${path}"
Following the naming conventions found inside
pass' code, "${passfile}" is the full path to the
`.gpg` file of the database, while "${path}" is the
former stripped of the "${PASSWORD_STORE_DIR}" ---
"${PREFIX}" inside the code --- value and of the
`.gpg` extension.

Change "${passfile}" to "${path}" and create a
variable with the name of the former corresponding
to the convention above.