Remove action options for -A/--action option
The user has four options to specify which action
is wanted: -p/--paste, -P/--paste-term, -t/--type
and -c/--copy. This makes necessary a case
statement for each one of them and also another
one for the prompt action.

Remove those options for a single one,
-A/--action, which takes as a parameter any of the
the long forms of the former without the double
dashes and the prompt action.

By doing this, not only the code base gets
smaller, but also it's simpler for the user to
remember one option and its parameters, then four
different options.