Remove find call for shopt -s globstar
To locate which password files were to be displayed
to the user with the xmenu, a `find` call inside a
subshell is used. As of its implementation, it's
very hard to read, since the string operations
occur inside of a `sh -c`, the `-exec` option
argument.

By using `shopt -s globstar`, an array is used and
the string operations (two string deletions) are
much more visible.