Fix test target
There are two problems with `test.bash` when it
comes to the expected result: it does not calls
the local copy of xmenu.bash in systems with the
extension globally installed, and, in systems
without it, it does not executes the local copy.

Fix the first problem by creating an unique name
for the local copy:
    > xmenu_$(date '+%F')_${hash_name:0:13}
in which `hash_name` is the output of `sha1sum` on
the `xmenu.bash` file.

Fix the second one by copying the file with
`install -m 0700` instead of `cp`.